+++ changelog.md
@@ -1,8 +1,12 @@
 # CHANGE LOG 
 ### How to use 
-See B0.01A, B = beta (version), 0 = major iteration, 01 = minor iteration, A = First in patch order
+See B-0.01.A, B = beta (version), 0 = major iteration, 01 = minor iteration, A = First in patch order
 
-## B0.01A
+## B-0.01.B
+- ### Added start to Expand Class
+- ### Added `HasChar(x, sp)` method
+
+## B-0.01.A
 - ### Added change log
 - ### Extended Constants Lib
     Added Constants.md (List of supported / planned lib constants)
@@ -17,3 +21,4 @@ See B0.01A, B = beta (version), 0 = major iteration, 01 = minor iteration, A = F
 
 - ### Added use of Log10 and not just LN
 - ### Re-Wrote Solve.java
+